Adding Homework_1
This homework is about Random Variate Generation. Inverse method of producing random number variate is used and for illustration, Geometric distribution Random variates have been demonstrated. A self-made function is defined to count the number of failures before a success. 

A simulation is demonstrated to verify the Central Limit Theorem and Law and Large Number.